concept-builder/requirements.txt
# OpenAI API
openai>=1.0.0

# Retry logic with exponential backoff
tenacity>=8.2.0

# Environment variable management
python-dotenv>=1.0.0

# YAML parsing
PyYAML>=6.0.0

# Image generation (optional - for social preview automation)
# Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
